+++ src/gfx.cpp
@@ -81,7 +81,7 @@ void R_Init() {
 
     GL_CALL(glClearColor, 0.2f, 0.3f, 0.3f, 1.0f);
     
-    for (int i = 0; i < MAX_LOCAL_CLIENTS; i++) {
+    for (size_t i = 0; i < MAX_LOCAL_CLIENTS; i++) {
         R_InitLocalClient(i);
         R_ClearTextDraws(i);
     }
@@ -191,7 +191,7 @@ void R_DrawFrame(int localClientNum) {
 void R_Frame() {
     RB_BeginFrame();
     GL_CALL(glEnable, GL_SCISSOR_TEST);
-    for (int i = 0; i < MAX_LOCAL_CLIENTS; i++) {
+    for (size_t i = 0; i < MAX_LOCAL_CLIENTS; i++) {
         if (!CG_LocalClientIsActive(i))
             continue;
 
@@ -202,7 +202,7 @@ void R_Frame() {
 }
 
 void R_WindowResized() {
-    for (int i = 0; i < MAX_LOCAL_CLIENTS; i++) {
+    for (size_t i = 0; i < MAX_LOCAL_CLIENTS; i++) {
         R_UpdateLocalClientView(i);
     }
 }
@@ -352,7 +352,7 @@ static void R_UnregisterDvars() {
 }
 
 void R_Shutdown() {
-    for(int i = 0; i < MAX_LOCAL_CLIENTS; i++)
+    for(size_t i = 0; i < MAX_LOCAL_CLIENTS; i++)
         R_ClearTextDraws(i);
 
     R_UnregisterDvars();
